Rearranging files as suggested by Graeme
